pyproject.toml
[project]
name = "ai-research-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.3",
    "aioredis>=2.0.1",
    "anthropic>=0.76.0",
    "fastmcp>=2.14.3",
    "httpx>=0.28.1",
    "kubernetes>=34.1.0",
    "mlflow>=3.8.1",
    "pydantic>=2.12.5",
    "pygithub>=2.8.1",
    "python-dotenv>=1.2.1",
    "redis>=7.1.0",
    "slack-sdk>=3.39.0",
    "toml>=0.10.2",
    "uvloop>=0.22.1",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.11",
    "types-redis>=4.6.0.20241004",
]
